+++ examples/tutorial/Main.hs
@@ -1,40 +1,23 @@
-{-# LANGUAGE TypeSynonymInstances           #-}
+{-|
+Module      : Main
+Description : Conway Game of Life example
+Copyright   : (c) Peter Lu, 2018
+License     : GPL-3
+Maintainer  : chippermonky@gmail.com
+Stability   : experimental
+-}
 
 module Main where
 
 import Smarties
 import System.Random
-import System.Console.Haskeline
-import System.Exit
 import Control.Concurrent
 import Control.Monad hiding (sequence)
-import Control.Applicative ((<$>))
-import Control.Monad.ST
-import Control.Lens
 import Prelude hiding (sequence)
 import Data.List
 import Data.List.Index (ifoldl)
-import Data.Maybe
 import qualified Data.Vector as V
-import qualified Data.Vector.Mutable as MV
 import Text.Printf
-
-type Pos = (Int, Int)
-
-addPos :: Pos -> Pos -> Pos
-addPos (x1,y1) (x2,y2) = (x1+y1, x2+y2)
-
-data Feelings = Happy | Sad | Hungry | Apathy deriving (Show, Eq)
-data Doings = BlowingBubbles | Eating | Bored deriving (Show)
-
-data Slime = Slime {
-    _pos :: Pos,
-    _feeling :: Feelings,
-    _doings :: Doings,
-    _weight :: Int
-} deriving (Show)
-
-makeLenses ''Slime
 
 -- world size parameters
 width :: Int
@@ -43,173 +26,128 @@ height :: Int
 height = 20
 numberCells :: Int
 numberCells = width * height
-neighbors :: [Pos]
-neighbors = [(-1,0),(1,0),(0,-1),(0,1)]
 
+-- types
+type Pos = (Int, Int)
+
+addPos :: Pos -> Pos -> Pos
+addPos (x1,y1) (x2,y2) = (x1+x2, y1+y2)
 
 wrapCoords :: Pos -> Pos
 wrapCoords (x,y) = (x `mod` width, y `mod` height)
 
--- TODO make sure this does what you want with neg coordinates...
+flattenCoords :: Pos -> Int
+flattenCoords (x,y) = y * width + x
+
 wrapFlattenCoords :: Pos -> Int
-wrapFlattenCoords (x,y) = (y `mod` height) * width + x `mod` width
+wrapFlattenCoords = flattenCoords . wrapCoords
+
+indexToPos :: Int -> Pos
+indexToPos p = (p `mod` width, p `div` width)
+
+neighbors :: [Pos]
+neighbors = [(-1,0),(1,0),(0,-1),(0,1),(1,1),(1,-1),(-1,1),(-1,-1)]
+
+countTrue :: [Bool] -> Int
+countTrue = foldl (\acc x -> if x then acc+1 else acc) 0
 
 -- behavior tree types
-type Slimes = [Slime]
-type SlimeGrid = V.Vector (Maybe Slime)
-type TreeStateType = (SlimeGrid, Slime)
-type ActionType = (Slime -> Slimes)
+type Grid = V.Vector Bool
+type TreeStateType = (Pos, Grid)
+type ActionType = Bool
 
--- | extract slime that is being operated on from behavior tree perception
-getMyself :: NodeSequence g TreeStateType ActionType Slime
-getMyself = do
-    (_, s) <- getPerception
-    return s
+-- behavior tree methods
+countNeighbors  :: NodeSequence g TreeStateType ActionType Int
+countNeighbors = do
+  (pos, grid) <- getPerception
+  return . countTrue . map ((grid V.!) . wrapFlattenCoords . addPos pos) $ neighbors
 
--- | extract a neighboring slime
-getSlimeRel :: Pos -> NodeSequence g TreeStateType ActionType (Maybe Slime)
-getSlimeRel p = do
-    (grid, _) <- getPerception
-    return $ grid V.! wrapFlattenCoords p
+ifNeighborsMoreThan :: Int -> NodeSequence g TreeStateType ActionType ()
+ifNeighborsMoreThan x = do
+  n <- countNeighbors
+  condition (n > x)
 
--- | get a list of neigboring slimes
-getNeighborSlimes :: NodeSequence g TreeStateType ActionType Slimes
-getNeighborSlimes = do
-    (grid, s) <- getPerception
-    return . mapMaybe ((grid V.!) . wrapFlattenCoords . addPos (_pos s)) $ neighbors
+ifNeighborsLessThan :: Int -> NodeSequence g TreeStateType ActionType ()
+ifNeighborsLessThan = flipResult . ifNeighborsMoreThan . (\x -> x-1)
 
--- behavior tree nodes
--- DELETE
-conditionSlimeIsFeeling :: Feelings -> Slime -> NodeSequence g TreeStateType ActionType ()
-conditionSlimeIsFeeling f s = fromCondition $
-    SimpleCondition (\_ -> _feeling s == f)
+die :: NodeSequence g TreeStateType ActionType ()
+die = fromAction $ SimpleAction (\_ -> False)
 
-actionMoveSlime :: Pos -> NodeSequence g TreeStateType ActionType ()
-actionMoveSlime p = fromAction $
-    SimpleAction (\_ -> \s -> [set pos (wrapCoords p) s])
--- |
-actionMoveSlimeRel :: Pos -> NodeSequence g TreeStateType ActionType ()
-actionMoveSlimeRel p = fromAction $
-    SimpleAction (\_ -> \s -> [over pos (wrapCoords . addPos p) s])
+born :: NodeSequence g TreeStateType ActionType ()
+born = fromAction $ SimpleAction (\_ -> True)
 
-actionSlime :: NodeSequence g TreeStateType ActionType ()
-actionSlime = actionCloneSlimeRel (0,0)
+ifAlive :: NodeSequence g TreeStateType ActionType ()
+ifAlive = do
+  (pos, grid) <- getPerception
+  condition $ grid V.! (wrapFlattenCoords pos)
 
-actionCloneSlimeRel :: Pos -> NodeSequence g TreeStateType ActionType ()
-actionCloneSlimeRel p = do
-    actionMoveSlimeRel p
-    actionSlime
-
+ifDead :: NodeSequence g TreeStateType ActionType ()
+ifDead = flipResult $ ifAlive
 
 -- | our behavior tree
-slimeTree :: (RandomGen g) => NodeSequence g TreeStateType ActionType ()
-slimeTree = do
-    nbs <- getNeighborSlimes
-    s <- getMyself
-    selector [
-        -- no neighbors
-        do
-            condition (length nbs == 0)
-            case _feeling s of
-                Happy -> actionMoveSlimeRel (1,0)
-                Sad -> actionMoveSlimeRel (-1,0)
-                Hungry -> actionMoveSlimeRel (0,1)
-                _ -> actionSlime
-        -- 1 neighbor
-        , do
-            condition (length nbs == 1)
-            s <- getMyself
-            let
-                nb = head nbs
-            case (_feeling nb, _feeling s) of
-                (Happy, Happy) -> actionMoveSlimeRel (0,1)
-                (Sad, Sad) -> actionMoveSlimeRel (0,-1)
-                (Sad, Happy) -> actionMoveSlime (_pos nb)
-                _ -> actionSlime
-        ]
+-- note rule 2. is a just a noop (Any live cell with two or three live neighbors lives on to the next generation.)
+conwayTree :: NodeSequence g TreeStateType ActionType ()
+conwayTree = do
+  selector [
+    -- rule 1. Any live cell with fewer than two live neighbors dies, as if by underpopulation.
+    do
+      ifAlive
+      ifNeighborsLessThan 2
+      die
+    -- rule 3. Any live cell with more than three live neighbors dies, as if by overpopulation.
+    , do
+      ifAlive
+      ifNeighborsMoreThan 3
+      die
+    -- rule 4. Any dead cell with exactly three live neighbors becomes a live cell, as if by reproduction.
+    , do
+      ifDead
+      ifNeighborsMoreThan 2
+      ifNeighborsLessThan 4
+      born]
 
-        -- > 1 neighbor case
-        -- don't do anything, this means the slime will die :(
+-- useful for testing :)
+potatoTree :: NodeSequence g TreeStateType ActionType ()
+potatoTree = do
+  selector [
+    do
+      ifDead
+      born
+    , do
+      ifAlive
+      die]
 
 
--- | DELETE
--- our slime action is a special case of behavior tree action type where there should only ever be one action in the output of the tree
--- we do a runtime check here to make sure this is the case
--- unfortunately smarties currently does not support type level checking of this constraint :(.
-extractHead :: [ActionType] -> ActionType
-extractHead fs
-    | null fs = (: [])
-    | length fs == 1 = head fs
-    | otherwise = error "slime behavior tree must only have one output"
-
--- | puts slimes in a grid
-makeSlimeGrid :: Slimes -> SlimeGrid
-makeSlimeGrid slimes = runST $ do
-    grid <- MV.replicate numberCells Nothing
-    forM_ slimes $ \s@(Slime (x,y) _ _ _) -> MV.write grid (y*width+x) (Just s)
-    V.freeze grid
-
--- | helper for writing slimes to console :)
-renderSlime :: Slime -> String
-renderSlime (Slime _ f _ _) = case f of
-    Happy -> "😊"
-    Sad -> "😟"
-    Hungry -> "😋"
-    Apathy -> "😐"
-
--- | helper for writing slimes to console :)
-renderSlimes :: Slimes -> String
-renderSlimes = Data.List.Index.ifoldl func "" . V.toList . makeSlimeGrid where
+-- | helper for writing grid to console :)
+renderGrid :: Grid -> String
+renderGrid = Data.List.Index.ifoldl func "" . V.toList where
     func acc i x = output where
-        nl = if i+1 `mod` width == 0 then "\n" else ""
+        nl = if (i+1) `mod` width == 0 then "\n" else ""
         se = case x of
-            Just s -> renderSlime s
-            Nothing -> " "
-        output = printf "%s %s %s" acc se nl
-
--- | fuse slimes that share the same cell
-fuseSlimes :: Slimes -> Slimes
-fuseSlimes slimes =  runST $ do
-    grid <- MV.replicate numberCells Nothing
-    forM_ slimes $ \(Slime (x,y) _ _ w) ->
-        MV.modify grid
-            (\case
-                Just (Slime _ _ _ w2) -> Just $ Slime (x,y) Happy Eating (w+w2)
-                Nothing -> Nothing)
-            (y*width+x)
-    catMaybes . V.toList <$> V.freeze grid
-
--- | run slimeTree for each slime collecting results
-slimeCycle :: (RandomGen g) => g -> Slimes -> (g, Slimes)
-slimeCycle g0 slimes = over _2 (fuseSlimes . concat) (mapAccumL runSlimeTree g0 slimes) where
-    -- function to run slime tree over all slimes accumulating the RNG
-    runSlimeTree g slime = (g', concat (map ($ slime) os)) where
-        (g', _, _, os) = execNodeSequence slimeTree g (makeSlimeGrid slimes, slime)
-
-applyNtimes :: (Num n, Ord n) => n -> (a -> a) -> a -> a
-applyNtimes 1 f x = f x
-applyNtimes n f x = f (applyNtimes (n-1) f x)
+            True -> "😱"
+            False -> "🌱"
+        output = printf "%s%s%s" acc se nl
 
 
-{-exitLoop :: IO ()
-exitLoop = do
-    minput <- getInputChar "% "
-    case minput of
-        Nothing -> return ()
-        Just 'q' -> exitSuccess-}
+-- | do conway rules for each grid cell
+-- TODO make this run concurrently :). Need to pregen list of generators.
+runConway :: g -> Grid -> (g, Grid)
+runConway g0 grid = (g', V.zipWith ($) (V.fromList fns) grid) where
+  (g', fns) = mapAccumL runTree g0 (map indexToPos [0..numberCells])
+  runTree g pos = (g'', if null os then id else \_ -> last os) where
+    (g'', _, _, os) = execNodeSequence conwayTree g (pos, grid)
 
+-- | go!
 main :: IO ()
 main = do
-    --forkIO exitLoop
     stdgen <- getStdGen
     let
-        genesis = [Slime (0,0) Sad Bored 1] -- 😢
-        --outSlimes = applyNtimes 100 (\(g,s) -> slimeCycle g s) (stdgen, genesis)
-        cycleOnce n (g,s) = do
-            let (g',s') = slimeCycle g s
-            putStrLn $ "gen " ++ show n
-            putStrLn $ renderSlimes s'
+        genesis = V.fromList . take numberCells $ randoms stdgen
+        cycleOnce (n :: Int) (g, s) = do
             putStrLn "done"
-            threadDelay 100
+            printf "gen %d\n" n
+            putStrLn $ renderGrid s
+            let (g',s') = runConway g s
+            threadDelay 100000
             cycleOnce (n+1) (g',s')
     cycleOnce 0 (stdgen, genesis)