+++ src/Smarties/Base.hs
@@ -3,7 +3,7 @@ Module      : Base
 Description : Base definitions and methods for Smarties
 Copyright   : (c) Peter Lu, 2018
 License     : GPL-3
-Maintainer  : chippermonky@email.com
+Maintainer  : chippermonky@gmail.com
 Stability   : experimental
 -}