+++ src/Smarties/Trans/Nodes.hs
@@ -3,7 +3,7 @@ Module      : Nodes
 Description : MTL equivalent of Smarties.Nodes
 Copyright   : (c) Peter Lu, 2018
 License     : GPL-3
-Maintainer  : chippermonky@email.com
+Maintainer  : chippermonky@gmail.com
 Stability   : experimental
 -}
 module Smarties.Trans.Nodes (